refactor(site): move analytics to root component
